[v2] refactor and create docs for results objects
This is the last thing needed for a documentation of the API.

https://embeddings-benchmark.github.io/mteb/

- renamed `load_results` to `results`
- moved `ModelResult` to its own script
- restructured import of results to use `results` module
- added documentation for the results module
- converted a few function marked with TODOs to private and deleted legacy loaders/converters
- moved `load_results.py` out of `results`
- Added missing documentation for a few of the types

sidenote: I really feel like we are starting to resolve some of our circular import issues, I rarely run into them now and when I do they are typically very easy to fix
