+++ docs/api/model.md
@@ -1,7 +1,5 @@
 # Models
 
-<!-- TODO: Encoder or model? Encoder is consistent with the code, but might be less used WDYT? We also use ModelMeta ... -->
-
 A model in `mteb` covers two concepts: metadata and implementation.
 - Metadata contains information about the model such as maximum input
 length, valid frameworks, license, and degree of openness.
