+++ mkdocs.yml
@@ -72,6 +72,7 @@ nav:
       - Benchmark: api/benchmark.md
       - Task: api/task.md
       - Model: api/model.md
+      - Results: api/results.md
       - Additional Types: api/types.md
   - Leaderboard: https://huggingface.co/spaces/mteb/leaderboard
