+++ mteb/__init__.py
@@ -5,12 +5,12 @@ from importlib.metadata import version
 from mteb.abstasks import AbsTask
 from mteb.abstasks.task_metadata import TaskMetadata
 from mteb.evaluate import evaluate
-from mteb.load_results import BenchmarkResults, load_results
-from mteb.load_results.task_results import TaskResult
+from mteb.load_results import load_results
 from mteb.models import Encoder, SentenceTransformerEncoderWrapper
 from mteb.models.get_model_meta import get_model, get_model_meta, get_model_metas
 from mteb.MTEB import MTEB
 from mteb.overview import TASKS_REGISTRY, get_task, get_tasks
+from mteb.results import BenchmarkResults, TaskResult
 
 from .benchmarks.benchmark import Benchmark
 from .benchmarks.get_benchmark import BENCHMARK_REGISTRY, get_benchmark, get_benchmarks
