+++ mteb/cache.py
@@ -11,9 +11,8 @@ from pathlib import Path
 from typing import cast
 
 from mteb.abstasks import AbsTask
-from mteb.load_results.benchmark_results import BenchmarkResults, ModelResult
-from mteb.load_results.task_results import TaskResult
 from mteb.models import ModelMeta
+from mteb.results import BenchmarkResults, ModelResult, TaskResult
 from mteb.types import ModelName, Revision
 
 logger = logging.getLogger(__name__)
