--- mteb/load_results/__init__.py
@@ -1,7 +1,0 @@
-from __future__ import annotations
-
-from .benchmark_results import BenchmarkResults, ModelResult
-from .load_results import load_results
-from .task_results import TaskResult
-
-__all__ = ["load_results", "TaskResult", "ModelResult", "BenchmarkResults"]
