+++ mteb/results/__init__.py
@@ -1,0 +1,6 @@
+from __future__ import annotations
+
+from .benchmark_results import BenchmarkResults, ModelResult
+from .task_result import TaskResult
+
+__all__ = ["TaskResult", "ModelResult", "BenchmarkResults"]
