+++ mteb/types/_result.py
@@ -2,14 +2,15 @@ from __future__ import annotations
 
 from typing import Any, NamedTuple
 
-HFSubset = (
-    str  # e.g. 'en-de', 'en', 'default' (default is used when there is no subset)
-)
-SplitName = str  # e.g. 'test', 'validation', 'train'
-Score = Any  # e.g. float, int, or any other type that represents a score, should be json serializable
-ScoresDict = dict[
-    str, Any
-]  # e.g {'main_score': 0.5, 'hf_subset': 'en-de', 'languages': ['eng-Latn', 'deu-Latn']}
+HFSubset = str
+"""The name of a HuggingFace dataset subset, e.g. 'en-de', 'en', 'default' (default is used when there is no subset)."""
+SplitName = str
+"""The name of a data split, e.g. 'test', 'validation', 'train'."""
+
+Score = Any
+"""A score value, could e.g. be accuracy. Normally it is a float or int, but it can take on any value. Should be json serializable."""
+ScoresDict = dict[str, Any]
+"""A dictionary of scores, typically also include metadata, e.g {'main_score': 0.5, 'accuracy': 0.5, 'f1': 0.6, 'hf_subset': 'en-de', 'languages': ['eng-Latn', 'deu-Latn']}"""
 
 
 class RetrievalEvaluationResult(NamedTuple):
