+++ tests/test_benchmark/mock_models.py
@@ -2,6 +2,7 @@
 
 from __future__ import annotations
 
+from types import SimpleNamespace as Namespace
 from typing import Any, Literal
 
 import numpy as np
@@ -12,7 +13,6 @@ from torch import Tensor
 from torch.utils.data import DataLoader
 
 from mteb.abstasks.task_metadata import TaskMetadata
-from mteb.load_results.task_results import Namespace
 from mteb.models.abs_encoder import AbsEncoder
 from mteb.models.model_meta import ModelMeta
 from mteb.models.sentence_transformer_wrapper import SentenceTransformerEncoderWrapper
